tests/wpt/tests/html/canvas/offscreen/canvas-context/2d.canvas.context.invalid.args.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.canvas.context.invalid.args</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.canvas.context.invalid.args</h1>
<p class="desc">Calling getContext with invalid arguments.</p>


<script>
var t = async_test("Calling getContext with invalid arguments.");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  assert_throws_js(TypeError, function() { canvas.getContext(''); });
  assert_throws_js(TypeError, function() { canvas.getContext('This is not an implemented context in any real browser'); });
  assert_throws_js(TypeError, function() { canvas.getContext('2d#'); });
  assert_throws_js(TypeError, function() { canvas.getContext('2d\0'); });
  assert_throws_js(TypeError, function() { canvas.getContext('2\uFF44'); });
  assert_throws_js(TypeError, function() { canvas.getContext('2D'); });
  assert_throws_js(TypeError, function() { canvas.getContext(); });
  assert_throws_js(TypeError, function() { canvas.getContext('null'); });
  assert_throws_js(TypeError, function() { canvas.getContext('undefined'); });
  t.done();

});
</script>
